+++ Packages/com.yontalane.interaction/CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 
-## 1.0.0 - 2022.06.01
+## 1.0.2 - 2022.06.01
+
+**Added**
+
+* Simple interaction has trigger strings to animate the interactable in addition to the interactor.
+
+**Changed**
+
+* Moved carry and simple interactions to Samples namespace.
+
+**Fixed**
+
+## 1.0.1 - 2022.06.01
 
 **Added**